increase timestep accuracy, bind timestep to duration of generalized time frequency, it vorks

+++ IEMath.h
@@ -314,6 +314,6 @@ inline bool BiasGreaterThan( real a, real b )
 
 const f32 gravityScale = 5.0f;
 const Vec2 gravity( 0, 10.0f * gravityScale );
-const float dt = 1.0f / 60.0f;
+const double dt = 1.0f / 60.0f;
 
 #endif // IEMATH_H
